Revert "workflows: add vcr to tests for the classifier"
This reverts commit f34cbeab3ec27657538977dd079fb870a763fac1.
